Renamed integration test
