+++ prices/src/test/java/com/challenge/prices/PricesApplicationTests.java
@@ -1,13 +1,30 @@
 package com.challenge.prices;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
 
-@SpringBootTest
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
 class PricesApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Autowired
+    private MockMvc mockMvc;
 
+    @Test
+    void testGetPriceById() throws Exception {
+        mockMvc.perform(get("/api/prices")
+                        .param("productId", "35455")
+                        .param("brandId", "1")
+                        .param("applicationDate", "2020-06-14T10:00:00"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.price").value(35.50))
+                .andExpect(jsonPath("$.currency").value("EUR"));
+    }
 }
